passed the syscall test case 'epoll' on Arm64 platform
EpollEvent in linux is not packed on Arm64.
Please see the linux code as reference:
In eventpoll.h
 #ifdef __x86_64__
 #define EPOLL_PACKED __attribute__((packed))
 #else
 #define EPOLL_PACKED
 #endif
